+++ resources/views/penelitian/penelitian_non_dilitabmas.blade.php
@@ -27,6 +27,8 @@
            <th>Kategori Bid</th>
            <th>Kategori TSE</th>
            <th>Tahun</th>
+           <th>Abstrak</th>
+
            <th></th>
            <th></th>
            <th></th>
@@ -53,49 +55,70 @@
           </td><td><strong data-toggle="tooltip" data-placement="top">
             {{$penelitian_non_dilitabmas->tahun}}</strong></td>
           </td>                
-          
+         <td>
 
+               <a class="btn btn-sm  glyphicon glyphicon-bookmark" data-toggle="modal" data-target="#abstrak-{{$penelitian_non_dilitabmas->id}}"></a>
+
+               <div class="modal fade" id="abstrak-{{$penelitian_non_dilitabmas->id}}" tabindex="-1" role="dialog" aria-labelledby="Title" aria-hidden="true">
+                <div class="modal-dialog" role="document">
+                  <div class="modal-content">
+                    <div class="modal-header">
+                      <strong><h3 class="modal-title" id="Title">Abstrak</h3></strong>
+                    </button>
+                  </div>
+                  <div class="modal-body">
+                   {{$penelitian_non_dilitabmas->abstrak}}
+                 </div>
+                 <div class="modal-footer">
+                  <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                </div>
+              </div>
+            </div>
+          </div>
+
+        </td>    
           @if(Session::get('userRole')!='Dosen')
           
           <td>
            <div class="dropdown">
             <a class="btn btn-sm btn-warning fa fa-plus-square" id="dropdownMenu1" data-toggle="dropdown" title="Tambahkan Luaran"></a>
             <ul class="dropdown-menu" aria-labelledby="dropdownMenu1">
-              <li ><a href="penelitian_non_dilitabmas/luaran_jurnal/{{$penelitian_non_dilitabmas->id}}">Jurnal</a></li>
-              <li class="divider"></li>
-              <li ><a href="penelitian_non_dilitabmas/luaran_buku_ajar/{{$penelitian_non_dilitabmas->id}}">Buku Ajar</a></li>
-              <li class="divider"></li>
-              <li ><a href="#">Pemakalah Forum Ilmiah</a></li>
-              <li class="divider"></li>
-              <li ><a href="#">HKI</a></li>
-              <li class="divider"></li>
-              <li ><a href="#">Lain-Lain</a></li>
-            </ul>
-          </div>
-        </td>
-        <td>
-          <a href="penelitian_non_dilitabmas/edit_non_dilitabmas/{{$penelitian_non_dilitabmas->id}}" class='btn btn-sm btn-info  fa fa-edit' title="Edit Penelitian"></a>
-        </td>
-
+             <li ><a href="penelitian_non_dilitabmas/luaran_buku_ajar/{{$penelitian_non_dilitabmas->id}}">Buku Ajar</a></li>
+             <li class="divider"></li>
+             <li ><a href="penelitian_non_dilitabmas/luaran_jurnal/{{$penelitian_non_dilitabmas->id}}">Jurnal</a></li>
+             
+             <li class="divider"></li>
+             <li ><a href="penelitian_non_dilitabmas/luaran_pemakalah/{{$penelitian_non_dilitabmas->id}}">Pemakalah Forum Ilmiah</a></li>
+         <li class="divider"></li>
+                    <li ><a href="penelitian_non_dilitabmas/luaran_hki/{{$penelitian_non_dilitabmas->id}}">HKI</a></li>
+                    <li class="divider"></li>
+                    <li ><a href="penelitian_non_dilitabmas/luaran_lain/{{$penelitian_non_dilitabmas->id}}">Lain-Lain</a></li>
+           </ul>
+         </div>
+       </td>
        <td>
-                <div>
-                  <a class="btn btn-sm btn-danger fa fa-trash" href="penelitian_non_dilitabmas/hapus_non_dilitabmas/{{$penelitian_non_dilitabmas->id}}" title="Hapus Penelitian"></a>
-                </div>              
-              </td>
-              
-          
-        @endif
-        
-      </tr>
-      @endforeach
-    </tbody>
-  </table>
+        <a href="penelitian_non_dilitabmas/edit_non_dilitabmas/{{$penelitian_non_dilitabmas->id}}" class='btn btn-sm btn-info  fa fa-edit' title="Edit Penelitian"></a>
+      </td>
+
+      <td>
+        <div>
+          <a class="btn btn-sm btn-danger fa fa-trash" onclick="return confirm('Anda Yakin Ingin Menghapus Data ?')" href="penelitian_non_dilitabmas/hapus_non_dilitabmas/{{$penelitian_non_dilitabmas->id}}" title="Hapus Penelitian"></a>
+        </div>              
+      </td>
+      
+      
+      @endif
+      
+    </tr>
+    @endforeach
+  </tbody>
+</table>
 </div>
 <div class="box-footer clearfix">
   @if(Session::get('userRole')!='Dosen')
   <a href="{{url('penelitian_non_dilitabmas/tambah_non_dilitabmas')}}" class="btn btn-md btn-success btn-flat pull-left"  style="margin-right:10px">Tambah</a>
 
-  <a href="{{url('penelitian/import')}}" class="btn btn-md btn-primary btn-flat pull-left">Ambil Data LITABMAS</a>
+  
 
   @endif
 </div>
