+++ resources/views/pengabdian/pengabdian_dilitabmas.blade.php
@@ -24,8 +24,9 @@
               <th>No</th>
               <th>Judul</th>
               <th>Ketua</th>
-              <th>Skema</th><th>Tahun</th>
-              
+              <th>Skema</th>
+              <th>Tahun</th>
+              <th>abstrak</th>
               @if(Session::get('userRole')!='Dosen')
               <th></th>
               <th></th>
@@ -50,7 +51,28 @@
                 </td><td><strong data-toggle="tooltip" data-placement="top">
                   {{$pengabdian_dilitabmas->tahun}}</strong></td>
                 </td>                
-                
+                   <td>
+
+               <a class="btn btn-sm  glyphicon glyphicon-bookmark" data-toggle="modal" data-target="#abstrak-{{$pengabdian_dilitabmas->id}}"></a>
+
+               <div class="modal fade" id="abstrak-{{$pengabdian_dilitabmas->id}}" tabindex="-1" role="dialog" aria-labelledby="Title" aria-hidden="true">
+                <div class="modal-dialog" role="document">
+                  <div class="modal-content">
+                    <div class="modal-header">
+                      <strong><h3 class="modal-title" id="Title">Abstrak</h3></strong>
+                    </button>
+                  </div>
+                  <div class="modal-body">
+                   {{$pengabdian_dilitabmas->abstrak}}
+                 </div>
+                 <div class="modal-footer">
+                  <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                </div>
+              </div>
+            </div>
+          </div>
+
+        </td>       
                 @if(Session::get('userRole')!='Dosen')
                 
                 <td>
@@ -60,9 +82,10 @@
                       
                     </a>
                     <ul class="dropdown-menu" aria-labelledby="dropdownMenu1">
-                      <li ><a href="pengabdian_dilitabmas/luaran_jurnal_peng/{{$pengabdian_dilitabmas->id}}">Jurnal</a></li>
-                      <li class="divider"></li>
                       <li ><a href="pengabdian_dilitabmas/luaran_buku_ajar_peng/{{$pengabdian_dilitabmas->id}}">Buku Ajar</a></li>
+                      <li class="divider"></li>
+                       <li ><a href="pengabdian_dilitabmas/luaran_jurnal_peng/{{$pengabdian_dilitabmas->id}}">Jurnal</a></li>
+                     
                       <li class="divider"></li>
                       <li ><a href="pengabdian_dilitabmas/luaran_pemakalah_peng/{{$pengabdian_dilitabmas->id}}">Pemakalah Forum Ilmiah</a></li>
                       <li class="divider"></li>
@@ -80,7 +103,7 @@
                 </td>
                 <td>
                   <div>
-                    <a class="btn btn-sm btn-danger fa fa-trash" href="pengabdian_dilitabmas/hapus_peng_dilitabmas/{{$pengabdian_dilitabmas->id}}" title="Hapus Penelitian"></a>
+                    <a class="btn btn-sm btn-danger fa fa-trash" onclick="return confirm('Anda Yakin Ingin Menghapus Data ?')" href="pengabdian_dilitabmas/hapus_peng_dilitabmas/{{$pengabdian_dilitabmas->id}}" title="Hapus Penelitian"></a>
                   </div>              
                 </td>
                 
@@ -95,8 +118,7 @@
           @if(Session::get('userRole')!='Dosen')
           <a href="{{url('pengabdian_dilitabmas/tambah_peng_dilitabmas')}}" class="btn btn-md btn-success btn-flat pull-left"  style="margin-right:10px">Tambah</a>
 
-          <a href="{{url('penelitian/import')}}" class="btn btn-md btn-primary btn-flat pull-left">Ambil Data LITABMAS</a>
-
+    
           @endif
         </div>
       </div>
