+++ resources/views/pengabdian/pengabdian_non_dilitabmas.blade.php
@@ -24,7 +24,9 @@
               <th>No</th>
               <th>Judul</th>
               <th>Ketua</th>
-              <th>Skema</th><th>Tahun</th>
+              <th>Skema</th>
+              <th>Tahun</th>
+              <th>Abstrak</th>
               @if(Session::get('userRole')!='Dosen')
               <th></th>
               <th></th>
@@ -51,7 +53,28 @@
                 </td><td><strong data-toggle="tooltip" data-placement="top">
                   {{$pengabdian_non_dilitabmas->tahun}}</strong></td>
                 </td>                
-                
+                <td>
+
+               <a class="btn btn-sm  glyphicon glyphicon-bookmark" data-toggle="modal" data-target="#abstrak-{{$pengabdian_non_dilitabmas->id}}"></a>
+
+               <div class="modal fade" id="abstrak-{{$pengabdian_non_dilitabmas->id}}" tabindex="-1" role="dialog" aria-labelledby="Title" aria-hidden="true">
+                <div class="modal-dialog" role="document">
+                  <div class="modal-content">
+                    <div class="modal-header">
+                      <strong><h3 class="modal-title" id="Title">Abstrak</h3></strong>
+                    </button>
+                  </div>
+                  <div class="modal-body">
+                   {{$pengabdian_non_dilitabmas->abstrak}}
+                 </div>
+                 <div class="modal-footer">
+                  <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                </div>
+              </div>
+            </div>
+          </div>
+
+        </td>    
                 @if(Session::get('userRole')!='Dosen')
                 
                 <td>
@@ -62,47 +85,48 @@
                       
                     </a>
                     <ul class="dropdown-menu" aria-labelledby="dropdownMenu1">
-                      <li ><a href="pengabdian_non_dilitabmas/luaran_jurnal_peng/{{$pengabdian_non_dilitabmas->id}}">Jurnal</a></li>
-                      <li class="divider"></li>
-                      <li ><a href="pengabdian_non_dilitabmas/luaran_buku_ajar_peng/{{$pengabdian_non_dilitabmas->id}}">Buku Ajar</a></li>
-                      <li class="divider"></li>
-                      <li ><a href="pengabdian_non_dilitabmas/luaran_pemakalah_peng/{{$pengabdian_non_dilitabmas->id}}">Pemakalah Forum Ilmiah</a></li>
-                      <li class="divider"></li>
-                      <li ><a href="pengabdian_non_dilitabmas/luaran_hki_peng/{{$pengabdian_non_dilitabmas->id}}">HKI</a></li>
-                      <li class="divider"></li>
-                      <li ><a href="pengabdian_non_dilitabmas/luaran_lain_peng/{{$pengabdian_non_dilitabmas->id}}">Lain-Lain</a></li>
-                    </ul>
-                  </div>
+                     <li ><a href="pengabdian_non_dilitabmas/luaran_buku_ajar_peng/{{$pengabdian_non_dilitabmas->id}}">Buku Ajar</a></li>
+                     <li class="divider"></li>
+                     <li ><a href="pengabdian_non_dilitabmas/luaran_jurnal_peng/{{$pengabdian_non_dilitabmas->id}}">Jurnal</a></li>
+                     
+                     <li class="divider"></li>
+                     <li ><a href="pengabdian_non_dilitabmas/luaran_pemakalah_peng/{{$pengabdian_non_dilitabmas->id}}">Pemakalah Forum Ilmiah</a></li>
+                     <li class="divider"></li>
+                     <li ><a href="pengabdian_non_dilitabmas/luaran_hki_peng/{{$pengabdian_non_dilitabmas->id}}">HKI</a></li>
+                     <li class="divider"></li>
+                     <li ><a href="pengabdian_non_dilitabmas/luaran_lain_peng/{{$pengabdian_non_dilitabmas->id}}">Lain-Lain</a></li>
+                   </ul>
+                 </div>
+               </td>
+               <td>
+                <a class="btn btn-sm btn-info  fa fa-edit" href="pengabdian_non_dilitabmas/edit_peng_non_dilitabmas/{{$pengabdian_non_dilitabmas->id}}" title="Edit Pengabdian"></a>
+                
+              </td>
+              <td>
+                <div>
+                  <a class="btn btn-sm btn-danger fa fa-trash" onclick="return confirm('Anda Yakin Ingin Menghapus Data ?')" href="pengabdian_non_dilitabmas/hapus_peng_non_dilitabmas/{{$pengabdian_non_dilitabmas->id}}"title=
+                    "Hapus Pengabdian"></a>
+                  </div>              
                 </td>
-                <td>
-                  <a class="btn btn-sm btn-info  fa fa-edit" href="pengabdian_non_dilitabmas/edit_peng_non_dilitabmas/{{$pengabdian_non_dilitabmas->id}}" title="Edit Pengabdian"></a>
-                  
-                </td>
-                <td>
-                  <div>
-                    <a class="btn btn-sm btn-danger fa fa-trash" href="pengabdian_non_dilitabmas/hapus_peng_non_dilitabmas/{{$pengabdian_non_dilitabmas->id}}"title=
-                      "Hapus Pengabdian"></a>
-                    </div>              
-                  </td>
-                  
-                  @endif
+                
+                @endif
 
-                </tr>
-                @endforeach
-              </tbody>
-            </table>
-          </div>
-          <div class="box-footer clearfix">
-            @if(Session::get('userRole')!='Dosen')
-            <a href="{{url('pengabdian_non_dilitabmas/tambah_peng_non_dilitabmas')}}" class="btn btn-md btn-success btn-flat pull-left"  style="margin-right:10px">Tambah</a>
-
-            <a href="{{url('penelitian/import')}}" class="btn btn-md btn-primary btn-flat pull-left">Ambil Data LITABMAS</a>
-
-            @endif
-          </div>
+              </tr>
+              @endforeach
+            </tbody>
+          </table>
         </div>
+        <div class="box-footer clearfix">
+          @if(Session::get('userRole')!='Dosen')
+          <a href="{{url('pengabdian_non_dilitabmas/tambah_peng_non_dilitabmas')}}" class="btn btn-md btn-success btn-flat pull-left"  style="margin-right:10px">Tambah</a>
 
+          
+
+          @endif
+        </div>
       </div>
-    </div><!-- /.row -->
-  </section><!-- /.content -->
-  @endsection
+
+    </div>
+  </div><!-- /.row -->
+</section><!-- /.content -->
+@endsection
